requirements.txt
torch>=1.4.0
torchvision>=0.5.0
yacs
numpy
scikit-image
Pillow
matplotlib
pyyaml
easydict
tqdm
efficientnet_pytorch
colorlog
tensorboard
opencv-python-headless